+++ src/JSystem/JKernel/JKRThread.cpp
@@ -82,7 +82,7 @@ void JKRThread::setCommon_heapSpecified(JKRHeap* heap, u32 stack_size, int param
     mStackMemory = JKRAllocFromHeap(mHeap, mStackSize, 0x20);
     mThreadRecord = (OSThread*)JKRAllocFromHeap(mHeap, sizeof(OSThread), 0x20);
 
-    void* stackBase = (void*)((int)mStackMemory + mStackSize);
+    void* stackBase = (void*)((intptr_t)mStackMemory + mStackSize);
     OSCreateThread(mThreadRecord, start, this, stackBase, mStackSize, param_3, 1);
 }
